modify no choose header bug

+++ Common/HFHttpRequest/HFHttpRequestParameters.m
@@ -94,7 +94,7 @@
 
 - (void)setInt:(NSInteger)value forKey:(NSString*)key
 {
-	[self setString:[NSString stringWithFormat:@"%d", value] forKey:key];
+	[self setString:[NSString stringWithFormat:@"%ld", (long)value] forKey:key];
 }
